Update and rename _item_sets_pve.go to item_sets_pve.go
Judgement Armor 8 piece set bonus might be wrong..
I have to add, that I can't test ANY of my commits.. I'm just hoping for the best XD
